Create extract_video_bios.sh
This script first gets the device address of the video card by running lspci and grepping for the VGA compatible controller. It then extracts the domain and bus number of the device from the device address.

Next, the script uses qemu-kvm to dump the video BIOS data from the device using PCI passthrough. The -device pci-assign,host=$DEVICE_ADDR option assigns the device to the guest, and the -dump-bios and -bios-size 64k options extract the BIOS data.

Finally, the script prints the extracted BIOS data using hexdump -C. You can save this script to a file, such as extract_video_bios.sh, and run it as a sudo user to extract the video BIOS data.