Create extractVBIOSfromVM.sh
To extract the video BIOS for a KVM virtual machine, you can use the following shell script:
Replace "your_vm_name" with the name of your virtual machine, and run the script as a user with sudo privileges.

The script uses the "virsh" command to get the ID of the virtual machine, and then uses "dumpxml" to extract the PCI address of the virtual graphics card. It then uses "qemu-img info" to extract the video BIOS from the virtual graphics card, and outputs the result to the console.